move sse2neon.h to extern
